Revert "Update gitignore and add workspace info"
This reverts commit 392786542f5f93d042bbabdfc9f0f6bfc6fda7b9. (+4 squashed commits)
Squashed commits:
[373b3c8] Revert "Update to recommended settings"

This reverts commit 1a5bad3ae64112acd081a737259ec9089213ae24.
[2b65692] Revert "Replace int with NSInteger"

This reverts commit 675ffcdf4975cce9d01126142a02eb567d23f095.
[144823f] Revert "Import AppKit rather than Cocoa"

This reverts commit 2f36b9aaae559a3d0c378c9b30a12ecbee2e6ccf.
[6b4b9da] Revert "Change bundle identifier prefix to match public location"

This reverts commit 29178f9e32fdf3686e057b59cc27f1302ef1fb13.

+++ src/Tomighty/Timer.h
@@ -14,7 +14,7 @@
 @interface Timer : NSObject
 
 - (id)initWithListener:(id <TimerListener>)listener;
-- (void)start:(NSInteger)minutes context:(TimerContext *)context;
+- (void)start:(int)minutes context:(TimerContext *)context;
 - (void)stop;
 
 @end
